refactor image pack rooms code to hook
